fix(core): disabling tests with skip
refs. metriport/metriport-internal#1931
