+++ packages/core/src/fhir-to-cda/cda-templates/components/__tests__/allergies.test.ts
@@ -25,7 +25,7 @@ beforeEach(() => {
   bundle = createEmptyBundle();
 });
 
-describe("buildAllergies", () => {
+describe.skip("buildAllergies", () => {
   it("returns an empty section if no resources found", () => {
     const res = buildAllergies(bundle);
     const cleanedJsonObj = removeEmptyFields(res);
