+++ packages/core/src/fhir-to-cda/cda-templates/components/__tests__/immunization.test.ts
@@ -25,7 +25,7 @@ beforeEach(() => {
   bundle = createEmptyBundle();
 });
 
-describe("buildImmunizations", () => {
+describe.skip("buildImmunizations", () => {
   it("correctly maps a single Immunization", () => {
     bundle.entry?.push({ resource: immunization });
     bundle.entry?.push({ resource: location });
