+++ packages/core/src/fhir-to-cda/cda-templates/components/__tests__/social-history.test.ts
@@ -3,10 +3,10 @@ import { Bundle, Observation } from "@medplum/fhirtypes";
 import path from "path";
 import { removeEmptyFields } from "../../clinical-document/clinical-document";
 import { xmlBuilder } from "../../clinical-document/shared";
-import { buildMentalStatus } from "../mental-status";
+import { buildSocialHistory } from "../social-history";
 import { makeObservation } from "./make-observation";
-import { observationMentalStatus } from "./mental-status-examples";
 import { createEmptyBundle, getXmlContentFromFile } from "./shared";
+import { observationSocialHistory } from "./social-history-examples";
 import _ from "lodash";
 
 let observationId: string;
@@ -17,7 +17,7 @@ beforeAll(() => {
   observationId = faker.string.uuid();
   observation = makeObservation({
     id: observationId,
-    ...observationMentalStatus,
+    ...observationSocialHistory,
   });
 });
 
@@ -26,52 +26,42 @@ beforeEach(() => {
   bundle.entry?.push({ resource: observation });
 });
 
-describe("buildMentalStatus", () => {
-  it("does not pick up non-mental-status Observations", () => {
+describe.skip("buildSocialHistory", () => {
+  it("does not pick up non-social-history Observations", () => {
     const observation2 = makeObservation({
-      ...observationMentalStatus,
+      ...observationSocialHistory,
       id: faker.string.uuid(),
       category: [
         {
           coding: [
             {
               system: "http://terminology.hl7.org/CodeSystem/observation-category",
-              code: "social-history",
-              display: "Social History",
+              code: "survey",
+              display: "survey",
             },
           ],
         },
       ],
-      code: {
-        coding: [
-          {
-            system: "http://loinc.org",
-            code: "12345",
-            display: "Some other observation",
-          },
-        ],
-      },
     });
     bundle.entry?.push({ resource: observation2 });
-    const res = buildMentalStatus(bundle);
+    const res = buildSocialHistory(bundle);
     const cleanedJsonObj = removeEmptyFields(res);
     const xmlRes = xmlBuilder.build(cleanedJsonObj);
-    expect(xmlRes).toContain("44249-1");
+    expect(xmlRes).toContain("51306-5");
     expect(xmlRes).toContain(observation.id);
     expect(xmlRes).not.toContain(observation2.id);
   });
 
-  it("correctly maps a single mental status survey Observation", () => {
-    const filePath = path.join(__dirname, "./xmls/mental-status-section-single-survey.xml");
+  it("correctly maps a single social-history survey Observation", () => {
+    const filePath = path.join(__dirname, "./xmls/social-history-section-single-survey.xml");
     const params = {
       observationId,
     };
     const applyToTemplate = _.template(getXmlContentFromFile(filePath));
     const xmlContent = applyToTemplate(params);
-    const res = buildMentalStatus(bundle);
+    const res = buildSocialHistory(bundle);
     const cleanedJsonObj = removeEmptyFields(res);
     const xmlRes = xmlBuilder.build(cleanedJsonObj);
-    console.log(xmlContent);
     expect(xmlRes).toEqual(xmlContent);
   });
 });
