+++ packages/core/src/fhir-to-cda/cda-templates/components/__tests__/vital-signs.test.ts
@@ -32,7 +32,7 @@ beforeEach(() => {
   bundle = createEmptyBundle();
 });
 
-describe("buildVitalSigns", () => {
+describe.skip("buildVitalSigns", () => {
   it("does not pick up non-vital-signs Observations", () => {
     bundle.entry?.push({ resource: observation });
     const observation2 = makeObservation({
